move file writers under _writers module
