+++ airbyte/_processors/sql/duckdb.py
@@ -15,7 +15,7 @@ from typing_extensions import Literal
 
 from airbyte._future_cdk import SqlProcessorBase
 from airbyte._future_cdk.sql_processor import SqlConfig
-from airbyte._processors.file import JsonlWriter
+from airbyte._writers.jsonl import JsonlWriter
 from airbyte.secrets.base import SecretString
