+++ airbyte/_processors/sql/postgres.py
@@ -8,7 +8,7 @@ from typing import Union
 from overrides import overrides
 
 from airbyte._future_cdk.sql_processor import SqlConfig, SqlProcessorBase
-from airbyte._processors.file import JsonlWriter
+from airbyte._writers.jsonl import JsonlWriter
 from airbyte.secrets.base import SecretString
