+++ airbyte/_processors/sql/snowflakecortex.py
@@ -24,7 +24,7 @@ if TYPE_CHECKING:
 
     from airbyte._future_cdk.catalog_providers import CatalogProvider
     from airbyte._future_cdk.state_writers import StateWriterBase
-    from airbyte._processors.file.base import FileWriterBase
+    from airbyte._writers.jsonl import FileWriterBase
 
 
 class SnowflakeCortexTypeConverter(SnowflakeTypeConverter):
