+++ airbyte/_writers/__init__.py
@@ -4,8 +4,7 @@
 from __future__ import annotations
 
 from airbyte._batch_handles import BatchHandle
-from airbyte._processors.file.base import FileWriterBase
-from airbyte._processors.file.jsonl import JsonlWriter
+from airbyte._writers.jsonl import FileWriterBase, JsonlWriter
 
 
 __all__ = [
