+++ airbyte/progress.py
@@ -43,10 +43,10 @@ if TYPE_CHECKING:
     from airbyte_protocol.models import AirbyteMessage
 
     from airbyte._message_iterators import AirbyteMessageIterator
+    from airbyte._writers.base import AirbyteWriterInterface
     from airbyte.caches.base import CacheBase
     from airbyte.destinations.base import Destination
     from airbyte.sources.base import Source
-    from airbyte.writers import AirbyteWriterInterface
 
 IS_REPL = hasattr(sys, "ps1")  # True if we're in a Python REPL, in which case we can use Rich.
 HORIZONTAL_LINE = "------------------------------------------------\n"
